requirements.txt
tensorflow-gpu==2.8.0
tqdm==4.64.1
Pillow==9.2.0
opencv-python==4.6.0.66
matplotlib==3.6.1